Clean up code and restructure project to seperate web, snake, and snake helpers.
