+++ Procfile
@@ -1,1 +1,1 @@
-web: gunicorn snake.wsgi --log-file -
+web: gunicorn wsgi --log-file -
