+++ snake/helpers/graph.py
