+++ urls.py
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.conf.urls import url
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
-from . import views
+import views
 
 urlpatterns = [
     url(r'^start', views.start),
