+++ views.py
@@ -4,7 +4,7 @@ from rest_framework.decorators import api_view
 from rest_framework.decorators import parser_classes
 from rest_framework.response import Response
 from rest_framework.parsers import JSONParser
-from .dva import DVA
+from snake.dva import DVA
 
 @api_view(['POST'])
 @parser_classes((JSONParser,))
