+++ wsgi.py
@@ -7,12 +7,8 @@ For more information on this file, see
 https://docs.djangoproject.com/en/1.10/howto/deployment/wsgi/
 """
 
-import os
-
 from django.core.wsgi import get_wsgi_application
 from whitenoise.django import DjangoWhiteNoise
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "snake.settings")
 
 application = get_wsgi_application()
 application = DjangoWhiteNoise(application)
